requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
gunicorn==21.2.0
uvicorn[standard]>=0.12.0 